+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ## Added
-- Assets now have a `srcSet` field available via GraphQL. ([#6660](https://github.com/craftcms/cms/issues/6660#issuecomment-714535759))
+- Assets now have a `srcset` field available via GraphQL. ([#6660](https://github.com/craftcms/cms/issues/6660))
 - The `restore` command now supports restoring zipped database backups. ([#7049](https://github.com/craftcms/cms/issues/7049))
 - Added `craft\services\Updates::getAreMigrationsPending()`. ([#7068](https://github.com/craftcms/cms/issues/7068))